requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
streamlit>=1.22.0
scikit-image>=0.19.0
matplotlib>=3.5.0 